Move ExportToXml private method to a public class
